Resolve environments before compiling sysimages
