--- sysimage-path.jl
@@ -1,23 +1,0 @@
-import SHA
-import Pkg
-
-"""
-    sysimage_path(dir)
-
-Return a suitable name for the system image of a project in `dir`. In order to
-avoid problems when/if Julia or the project dependencies get updated, all
-versions get encoded in the system image file name.
-"""
-function sysimage_path(dir)
-    ext = if Sys.iswindows()
-        "dll"
-    elseif Sys.isapple()
-        "dylib"
-    else
-        "so"
-    end
-
-    hash = open(SHA.sha1, joinpath(dir, "Manifest.toml")) |> bytes2hex
-
-    joinpath(dir, "eglot-jl_$(Base.VERSION)_$hash.$ext")
-end
